\name{scale_colour_helse}
\alias{scale_colour_helse}
\alias{scale_color_helse}
\alias{scale_fill_helse}
\title{ggplot2 colour and fill scales for Norwegian Health Trust colours}
\usage{
scale_colour_helse(..., reverse = FALSE)

scale_color_helse(..., reverse = FALSE)

scale_fill_helse(..., reverse = FALSE)
}
\arguments{
\item{...}{Additional arguments passed to the ggplot2 scale.}

\item{reverse}{If \code{TRUE}, reverse the default colour order.}
}
\value{
A ggplot2 scale.
}
\description{
Discrete ggplot2 colour and fill scales using the Norwegian Health Trust
palette.
}
